backend/requirements.txt
Django==3.1
djangorestframework==3.11.1
psycopg2==2.8.5
requests==2.24.0
pendulum==2.1.2